add CP_idr and DP_idr functions
