+++ idrstream_dp_env.yml
@@ -1,4 +1,4 @@
-name: idrstream
+name: idrstream_dp
 channels:
   - conda-forge
 dependencies:
